requirements.txt
mysql-connector-python
python-dotenv
plyer
concurrent-log-handler
pywhatkit
pyautogui